// src/types/Note.ts
export enum Note {
  A = "A",
  AFlat = "Ab",
  ASharp = "A#",
  B = "B",
  BFlat = "Bb",
  BSharp = "B#",
  C = "C",
  CFlat = "Cb",
  CSharp = "C#",
  D = "D",
  DFlat = "Db",
  DSharp = "D#",
  E = "E",
  EFlat = "Eb",
  ESharp = "E#",
  F = "F",
  FFlat = "Fb",
  FSharp = "F#",
  G = "G",
  GFlat = "Gb",
  GSharp = "G#",
}